pyproject.toml
[project]
name = "ghostlab-analytics-models"
version = "0.1.0"
description = "Local-first analytics and ML pipelines for Ghost Laboratory"
requires-python = ">=3.10"

dependencies = [
  "numpy",
  "pandas",
  "pyarrow",
  "scikit-learn",
  "torch",
  "python-dotenv",
  "rich",
  "scipy",
  "google-cloud-bigquery",
  "db-dtypes",
]

[project.scripts]
ghostlab = "ghostlab.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
